+++ CharacterMap/CharacterMap/Services/GlyphService.cs
@@ -82,15 +82,15 @@ namespace CharacterMap.Services
             return Task.CompletedTask;
         }
 
-        internal static string GetCharacterDescription(int unicodeIndex, FontVariant variant)
+        internal static string GetCharacterDescription(uint unicodeIndex, FontVariant variant)
         {
             if (variant == null || _provider == null)
                 return null;
 
-            return _provider.GetCharacterDescription(unicodeIndex, variant);
+            return _provider.GetCharacterDescription((int)unicodeIndex, variant);
         }
 
-        internal static string GetCharacterKeystroke(int unicodeIndex)
+        internal static string GetCharacterKeystroke(uint unicodeIndex)
         {
             if (unicodeIndex >= 128 && unicodeIndex <= 255)
                 return Localization.Get("CharacterKeystrokeLabel",  unicodeIndex);
